Update README.md
This ML project is to analyzes the sentiments shown in the tweets by automatic learning over 1,600,000 tweets.

Sentiment analysis refers to identifying and classifying them to positive or negative accordingly, which is expressed in text sources. These data are so useful in predicting the opinion of the majority of people over any sensitive issue.